pyproject.toml
[project]
name = "regression-monkey-nonparametric"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "RyoNak", email = "nakagamiryo0901@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "scipy>=1.16.1",
]

[project.scripts]
regression-monkey-nonparametric = "regression_monkey_nonparametric:main"


[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
book = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "nbformat>=5.10.4",
    "pandas>=2.3.2",
    "pyyaml>=6.0.2",
    "plotly>=6.3.0",
    "nbclient>=0.10.2",
    "google-cloud-bigquery>=3.36.0",
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery-storage>=2.33.0",
]
dev = [
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.10",
]
